requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.5.2
asttokens==2.1.0
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
Django==4.1.3
django-clearcache==1.2.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.7.0
django-extensions==3.2.1
django-filter==22.1
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
drf-yasg==1.21.4
executing==1.2.0
Faker==17.6.0
flake8==6.0.0
flake8-quotes==3.3.2
idna==3.4
inflection==0.5.1
ipython==8.6.0
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mccabe==0.7.0
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
prompt-toolkit==3.0.32
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.1
traitlets==5.5.0
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
